backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
python-decouple==3.8
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
python-docx==1.1.0
WeasyPrint==60.1
reportlab==4.0.7
boto3==1.29.7
django-storages==1.14.2
pytest==7.4.3
pytest-django==4.7.0